requirements.txt
# ---- Web Framework ----
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
python-multipart>=0.0.20,<1.0.0

# ---- LangChain + Groq LLM ----
langchain-core>=0.3.0,<1.0.0
langchain-openai>=0.3.0,<1.0.0

# ---- Speech (STT / TTS) ----
groq>=0.25.0,<1.0.0
edge-tts>=7.2.7

# ---- MongoDB ----
motor>=3.6.0,<4.0.0
pymongo>=4.10.0,<5.0.0

# ---- RAG (Reading Module) ----
langchain-mongodb
pillow
duckduckgo-search
